+++ R/plot_precision_recall.R
@@ -70,58 +70,20 @@ plot_precision_recall <- function(peakfiles,
                                  cast = TRUE,
                                  save_path = save_path)
     #### Plot precision-recall ####
-    gg <- ggplot2::ggplot(
-        data = plot_dat, 
-        ggplot2::aes_string(x="recall", 
-                            y="precision",
-                            group="peaklist1", 
-                            color=color)) + 
-        ggplot2::geom_point(ggplot2::aes_string(size = "1-threshold",
-                                                shape=shape),
-                            alpha=.8) +
-        ggplot2::geom_line() + 
-        ggplot2::facet_grid(facets = facets) + 
-        ggplot2::ylim(0, 100) +
-        ggplot2::xlim(0, 100) +
-        ggplot2::labs(
-            title="precision-recall curves", 
-            subtitle = subtitle,
-            x="Recall\n(% reference peaks in sample peaks)", 
-            y="Precision\n(% sample peaks in reference peaks)") + 
-        ggplot2::theme_bw() +
-        ggplot2::theme(
-            strip.background = ggplot2::element_rect(fill = "grey20"),
-            strip.text = ggplot2::element_text(color="white"),
-            plot.margin = ggplot2::margin(.5,.5,.5,.5)
-            # legend.title = ggplot2::element_text(size=7),
-            # legend.spacing.y = ggplot2::unit(.001, units = "npc"),
-            # legend.text=ggplot2::element_text(size=7)
-            ) 
-    if(interact) gg <- plotly::ggplotly(gg)
+    gg <- plot_precision_recall_prcurve(plot_dat=plot_dat,
+                                        color=color,
+                                        shape=shape,
+                                        facets=facets,
+                                        subtitle=subtitle, 
+                                        interact=interact)
     #### Plot F1 #####
-    if(plot_f1){
-        ggf1 <- ggplot2::ggplot(
-            data = plot_dat, 
-            ggplot2::aes_string(x="threshold", 
-                                y="F1",
-                                group="peaklist1", 
-                                color=color)) + 
-            ggplot2::geom_point(ggplot2::aes_string(size = "1-threshold",
-                                                    shape=color),
-                                alpha=.8) +
-            ggplot2::facet_grid(facets = facets) + 
-            ggplot2::ylim(0, 100) +
-            ggplot2::xlim(initial_threshold, 1) +
-            ggplot2::labs(
-                title="F1 plot",
-                y="F1\n2 * (precision * recall) / (precision + recall)") + 
-            ggplot2::geom_line() +
-            ggplot2::theme_bw() +
-            ggplot2::theme(
-                strip.background = ggplot2::element_rect(fill = "grey20"),
-                strip.text = ggplot2::element_text(color="white")
-            )
-        if(interact) ggf1 <- plotly::ggplotly(ggf1)
+    if(isTRUE(plot_f1)){
+        ggf1 <- plot_precision_recall_f1(plot_dat=plot_dat,
+                                         color=color,
+                                         shape=shape,
+                                         facets=facets,
+                                         initial_threshold=initial_threshold,
+                                         interact=interact)
     } else {
         ggf1 <- NULL
     }